+++ packages/patrol_cli/lib/src/commands/build_android.dart
@@ -135,6 +135,7 @@ class BuildAndroidCommand extends PatrolCommand {
 
     final displayLabel = boolArg('label');
     final uninstall = boolArg('uninstall');
+    final noTreeShakeIcons = boolArg('no-tree-shake-icons');
 
     final customDartDefines = {
       ..._dartDefinesReader.fromFile(),
@@ -187,6 +188,7 @@ class BuildAndroidCommand extends PatrolCommand {
       appServerPort: super.appServerPort,
       testServerPort: super.testServerPort,
       uninstall: uninstall,
+      noTreeShakeIcons: noTreeShakeIcons,
     );
 
     try {