+++ packages/patrol_cli/CHANGELOG.md
@@ -7,6 +7,7 @@
 
 - Add help command `h` to `patrol develop`. (#2701)
 - Add printing paths to the APKs after `patrol build` command. (#2685)
+- Add support for `--no-tree-shake-icons` flag in cli commands (#2704)
 
 ## 3.9.0